+++ tests/Functional/Model/FullDocument.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * This file is part of the Sulu CMS.
+ * This file is part of Sulu.
  *
  * (c) MASSIVE ART WebServices GmbH
  *
@@ -11,8 +11,7 @@
 
 namespace Sulu\Component\DocumentManager\Tests\Functional\Model;
 
-use Sulu\Component\DocumentManager\Behavior\Audit\BlameBehavior;
-use Sulu\Component\DocumentManager\Behavior\Audit\TimestampBehavior;
+use Sulu\Component\DocumentManager\Behavior\Audit\LocalizedTimestampBehavior;
 use Sulu\Component\DocumentManager\Behavior\Mapping\ChildrenBehavior;
 use Sulu\Component\DocumentManager\Behavior\Mapping\LocaleBehavior;
 use Sulu\Component\DocumentManager\Behavior\Mapping\NodeNameBehavior;
@@ -25,8 +24,7 @@ use Sulu\Component\DocumentManager\Behavior\Mapping\UuidBehavior;
  */
 class FullDocument implements
     NodeNameBehavior,
-    TimestampBehavior,
-    BlameBehavior,
+    LocalizedTimestampBehavior,
     ParentBehavior,
     UuidBehavior,
     ChildrenBehavior,