+++ tests/Unit/PathSegmentRegistryTest.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * This file is part of the Sulu CMS.
+ * This file is part of Sulu.
  *
  * (c) MASSIVE ART WebServices GmbH
  *
@@ -15,12 +15,19 @@ use Sulu\Component\DocumentManager\PathSegmentRegistry;
 
 class PathSegmentRegistryTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var PathSegmentRegistry
+     */
+    private $pathRegistry;
+
     public function setUp()
     {
-        $this->pathRegistry = new PathSegmentRegistry([
-            'base' => 'cmf',
-            'foobar' => 'barfoo',
-        ]);
+        $this->pathRegistry = new PathSegmentRegistry(
+            [
+                'base' => 'cmf',
+                'foobar' => 'barfoo',
+            ]
+        );
     }
 
     /**